// src/utils/interfaces/IAddressViaCEP.ts
export interface IAddressViaCEP {
   data:{
      cep:         string;
      logradouro:  string;
      complemento: string;
      bairro:      string;
      localidade:  string;
      uf:          string;
      ibge:        string;
      gia:         string;
      ddd:         string;
      siafi:       string;
   }
}